+++ src/main/java/it/com/acamir/veicolibe/repository/VeicoloRepository.java
@@ -15,6 +15,6 @@ public interface VeicoloRepository extends JpaRepository<Veicolo, Integer>, JpaS
  @Query(value = "select * from veicolo a"
     		+ " where (?1 = '' or a.matricola = ?1) and (?2 = '' or a.telaio = ?2) "
     		+ " and (?3 is null or a.assegnatario = ?3) order by a.matricola" , nativeQuery = true)
- List<Veicolo> getListVeicoloByFilter(String matricola, String telaio, Integer idAssegnatazio);
+ List<Veicolo> getListVeicoloByFilter(String matricola, String telaio, Long idAssegnatazio);
 	
 }
